requirements.txt
agithub==2.2.2
alabaster==0.7.12
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==5.0.7
docutils==0.16
idna==2.10
imagesize==1.2.0
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
numpy==1.20.2
packaging==20.9
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pyzmq==22.0.3
requests==2.25.1
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.0.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
